+++ src/client/vite-env.d.ts
@@ -1,5 +1,7 @@
 /// <reference types="vite/client" />
 
+declare const __APP_GIT_SHA__: string;
+
 declare module "guacamole-common-js" {
   const Guacamole: any;
   export default Guacamole;